tools: move `RepoLike` to tools
